Add tests and income/expenses view

+++ src/main/java/net/ivan/kavaliou/moneyman/model/persistence/Transaction.java
@@ -12,6 +12,7 @@ import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -39,7 +40,7 @@ public class Transaction  {
 
     @NotNull
     @NotBlank
-    private  float value;
+    private BigDecimal value;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "id_transaction_category", nullable = false)
